ai-api/requirements.txt
fastapi
uvicorn
faster-whisper
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
transformers
open-clip-torch
opencv-python-headless
numpy
Pillow
python-multipart
